requirements.txt
annotated-types==0.7.0
anyio==4.12.0
boto3==1.28.14
botocore==1.31.85
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.115.8
fastapi-cli==0.0.7
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
jmespath==1.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
passlib==1.7.4
psycopg2==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
alembic==1.11.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.0
s3transfer==0.6.2
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.38
starlette==0.45.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.0.7
uvicorn==0.34.0
watchfiles==1.1.1
websockets==15.0.1
